refactored internet connection check and fixed #158
