+++ app/src/main/java/ro/code4/monitorizarevot/ui/base/BaseFragment.kt
@@ -27,21 +27,4 @@ abstract class BaseFragment<out T : BaseViewModel> : Fragment(), Layout,
     ): View? {
         return inflater.inflate(layout, container, false)
     }
-
-    fun isOnline(context: Context): Boolean {
-        val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            val n = cm.activeNetwork
-            if (n != null) {
-                val nc = cm.getNetworkCapabilities(n)
-                //It will check for both wifi and cellular network
-                return nc!!.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) || nc.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
-            }
-            return false
-        } else {
-            val netInfo = cm.activeNetworkInfo
-            return netInfo != null && netInfo.isConnectedOrConnecting
-        }
-    }
 }
